Compilation without warnings
Former-commit-id: 15f292028e8d202a78efd9a1eaf28fedb70d8b89

+++ libraries/parallel/xmipp_mpi.h
@@ -48,7 +48,7 @@ class MpiNode
 public:
 
     //MPI_Comm *comm;
-    int rank, size, active;//, activeNodes;
+    size_t rank, size, active;//, activeNodes;
     MpiNode(int &argc, char ** argv);
     ~MpiNode();
